requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.14.0=pypi_0
alabaster=0.7.12=pypi_0
anyqt=0.0.13=pyhd3eb1b0_0
arff=0.9=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
attrs=21.4.0=pypi_0
audioread=3.0.1=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pypi_0
baycomp=1.0.2=py_1
blas=1.0=mkl
bleach=4.1.0=pypi_0
bottleneck=1.3.2=py37heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachecontrol=0.12.6=pyhd3eb1b0_0
cached-property=1.5.2=pypi_0
certifi=2021.10.8=py37h06a4308_0
cffi=1.15.0=py37h7f8727e_0
chardet=4.0.0=py37h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=pypi_0
commonmark=0.9.1=pyhd3eb1b0_0
cryptography=36.0.0=py37h9ce1e76_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.28=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dictdiffer=0.8.0=py_0
docutils=0.17.1=pypi_0
entrypoints=0.4=pypi_0
et_xmlfile=1.1.0=py37h06a4308_0
expat=2.4.1=h2531618_2
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.31.2=pypi_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py37_1
gast=0.5.2=pypi_0
gensim=4.1.2=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.63.1=h5a9c865_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-pasta=0.2.0=pypi_0
grpcio=1.40.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h11=0.12.0=pyhd3eb1b0_0
h2=3.2.0=py37_1
h5py=3.4.0=pypi_0
himalaya=0.3.5=pypi_0
hpack=3.0.0=py37h28b3542_0
html5lib=1.1=pypi_0
httpcore=0.13.3=pyhd3eb1b0_0
httpx=0.18.2=pyhd3eb1b0_0
hyperframe=5.2.0=py_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imageio=2.16.1=pypi_0
imageio-ffmpeg=0.4.5=pypi_0
imagesize=1.3.0=pypi_0
importlib-metadata=4.8.2=py37h06a4308_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.0=pypi_0
ipython=7.31.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jeepney=0.7.1=pyhd3eb1b0_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
keyring=23.4.0=py37h06a4308_0
keyrings.alt=4.1.0=py37h06a4308_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libllvm11=11.1.0=h3826bc1_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
librosa=0.10.1=pypi_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.37.0=py37h295c915_1
lockfile=0.12.2=py37h06a4308_0
lxml=4.8.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.5.0=py37h06a4308_0
matplotlib-base=3.5.0=py37h3ed280b_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
msgpack-python=1.0.2=py37hff7bd54_1
munkres=1.1.4=py_0
nbclient=0.5.10=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
nibabel=3.2.2=pypi_0
ninja=1.10.2=py37hd09550d_3
nltk=3.6.2=pyhd3eb1b0_0
notebook=6.4.8=pypi_0
numba=0.54.1=py37h51133e4_0
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
numpydoc=1.2=pypi_0
olefile=0.46=py37_0
opencv-python=4.5.5.64=pypi_0
openh264=2.1.0=hd408876_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.0.2u=h7b6447c_0
opentsne=0.4.3=py37hfd86e86_1
orange-canvas-core=0.1.15=pyhd3eb1b0_0
orange-widget-base=4.7.0=py_1
orange3=3.26.0=py37he6710b0_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.2=pypi_0
pip=21.2.2=py37h06a4308_0
platformdirs=4.0.0=pypi_0
pooch=1.8.1=pypi_0
pot=0.9.0=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.28=pypi_0
protobuf=3.18.0=pypi_0
ptyprocess=0.7.0=pypi_0
pycortex=1.3.0.dev0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pygments=2.11.2=pypi_0
pynndescent=0.5.4=pyhd3eb1b0_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyqtgraph=0.11.0=py_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py37_1
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
pytorch=1.8.1=py3.7_cuda10.1_cudnn7.6.3_0
pytz=2021.1=pypi_0
pyyaml=6.0=py37h7f8727e_1
pyzmq=22.3.0=pypi_0
qt=5.9.6=h8703b6f_2
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
readline=7.0=h7b6447c_5
regex=2021.8.3=py37h7f8727e_0
requests=2.26.0=pyhd3eb1b0_0
rfc3986=1.4.0=pyhd3eb1b0_0
scikit-learn=1.0=pypi_0
scikit-multilearn=0.2.0=pypi_0
scipy=1.2.0=pypi_0
seaborn=0.11.2=pypi_0
secretstorage=3.3.1=py37h06a4308_0
send2trash=1.8.0=pypi_0
serverfiles=0.3.1=py37h06a4308_0
setuptools=58.0.4=py37h06a4308_0
shapely=1.8.1.post1=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
smart-open=5.2.1=pypi_0
sniffio=1.2.0=py37h06a4308_1
snowballstemmer=2.2.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
sphinx=4.4.0=pypi_0
sphinx-gallery=0.10.1=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.33.0=h62c20be_0
tbb=2021.4.0=hd09550d_0
tensorboard=1.14.0=pypi_0
tensorboardx=2.4=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.8.1=py37
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchstat=0.0.7=pypi_0
torchvision=0.9.1=py37_cu101
tornado=6.1=py37h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=1.26.7=pyhd3eb1b0_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.1=pypi_0
werkzeug=2.0.1=pypi_0
wget=3.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
wrapt=1.12.1=pypi_0
x264=1!157.20191217=h7b6447c_0
xlrd=2.0.1=pyhd3eb1b0_0
xlsxwriter=3.0.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
